rewrote base function parts nicer
instead of having a dict of basefunctions that is loaded and kept
separately in the widgets, we now use a class to hold all the metadata
of a base function. This is similar to the way it is implemenet in the
OpenSCAD backend. The module with the function in it is only loaded on
demand.