workaround SERVER_NAME
to use url_for in absence of a request, Flask requires a SERVER_NAME. As
this turns out to be problematic for OpenShift, we call execute url_for
only later, when a request is around. For this the parameters for the
call need to be stored in the index. For serialization we use unicode
and eval. As the is not user submitted data this is save.